refactor(workflow): isolate bounded collection runtime
